pyproject.toml
[tool.poetry]
name = "avito_announcment"
version = "0.1.0"
description = ""
authors = ["ssyucfa <patin123321@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
Pillow = "^9.1.0"
dj-rest-auth = "^2.2.4"
django-allauth = "^0.50.0"
djangorestframework-simplejwt = "^5.1.0"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
